+++ nevergrad/common/typing.py
@@ -45,11 +45,11 @@ import numpy as _np
 
 
 ArgsKwargs = Tuple[Tuple[Any, ...], Dict[str, Any]]
-ArrayLike = Union[Tuple[float, ...], List[float], _np.ndarray]
+ArrayLike = Union[Tuple[float, ...], List[float], _np.ndarray]  # most common
 PathLike = Union[str, Path]
 FloatLoss = float
 Loss = Union[float, ArrayLike]
-BoundValue = Optional[Union[float, int, _np.int_, _np.float_, _np.ndarray]]
+BoundValue = Optional[Union[float, int, _np.int_, _np.float_, ArrayLike]]
 
 
 # %% Protocol definitions for executor typing