+++ nevergrad/optimization/differentialevolution.py
@@ -20,13 +20,15 @@ class Crossover:
             self.CR = crossover
         elif crossover == "random":
             self.CR = self.random_state.uniform(0.0, 1.0)
-        elif crossover not in ["twopoints", "onepoint"]:
+        elif crossover not in ["twopoints", "onepoint", "rotated_twopoints"]:
             raise ValueError(f'Unknown crossover "{crossover}"')
 
     def apply(self, donor: np.ndarray, individual: np.ndarray) -> None:
         dim = donor.size
         if self.crossover == "twopoints" and dim >= 4:
             return self.twopoints(donor, individual)
+        elif self.crossover == "rotated_twopoints" and dim >= 4:
+            return self.rotated_twopoints(donor, individual)
         elif self.crossover == "onepoint" and dim >= 3:
             return self.onepoint(donor, individual)
         else:
@@ -56,6 +58,15 @@ class Crossover:
         else:
             donor[: bounds[0]] = individual[: bounds[0]]
             donor[bounds[1] :] = individual[bounds[1] :]
+
+    def rotated_twopoints(self, donor: np.ndarray, individual: np.ndarray) -> None:
+        bounds = sorted(self.random_state.choice(donor.size + 1, size=2, replace=False).tolist())
+        if bounds[1] == donor.size and not bounds[0]:  # make sure there is at least one point crossover
+            bounds[self.random_state.randint(2)] = self.random_state.randint(1, donor.size)
+        bounds2 = [self.random_state.choice(donor.size + 1 - bounds[1] + bounds[0])]
+        bounds2.append(bounds2[0] + bounds[1] - bounds[0])
+        assert bounds[1] < donor.size + 1
+        donor[bounds[0] : bounds[1]] = individual[bounds2[0] : bounds2[1]]
 
 
 class _DE(base.Optimizer):
@@ -100,6 +111,7 @@ class _DE(base.Optimizer):
         self._uid_queue = base.utils.UidQueue()
         self.population: tp.Dict[str, p.Parameter] = {}
         self.sampler: tp.Optional[base.Optimizer] = None
+        self._no_hypervolume = self._config.multiobjective_adaptation
 
     def recommend(self) -> p.Parameter:  # This is NOT the naive version. We deal with noise.
         if self._config.recommendation != "noisy":
@@ -248,6 +260,7 @@ class DifferentialEvolution(base.ConfiguredOptimizer):
         - "random": different random (uniform) crossover rate at each iteration
         - "onepoint": one point crossover
         - "twopoints": two points crossover
+        - "rotated_twopoints": more genetic 2p cross-over
         - "parametrization": use the parametrization recombine method
     F1: float
         differential weight #1
@@ -283,6 +296,7 @@ class DifferentialEvolution(base.ConfiguredOptimizer):
         assert isinstance(crossover, float) or crossover in [
             "onepoint",
             "twopoints",
+            "rotated_twopoints",
             "dimension",
             "random",
             "parametrization",
@@ -300,6 +314,10 @@ class DifferentialEvolution(base.ConfiguredOptimizer):
 
 DE = DifferentialEvolution().set_name("DE", register=True)
 TwoPointsDE = DifferentialEvolution(crossover="twopoints").set_name("TwoPointsDE", register=True)
+RotatedTwoPointsDE = DifferentialEvolution(crossover="rotated_twopoints").set_name(
+    "RotatedTwoPointsDE", register=True
+)
+
 LhsDE = DifferentialEvolution(initialization="LHS").set_name("LhsDE", register=True)
 QrDE = DifferentialEvolution(initialization="QR").set_name("QrDE", register=True)
 NoisyDE = DifferentialEvolution(recommendation="noisy").set_name("NoisyDE", register=True)