+++ nevergrad/optimization/test_optimizerlib.py
@@ -117,6 +117,10 @@ def check_optimizer(
 SLOW = [
     "NoisyDE",
     "NoisyBandit",
+    "Noisy13Splits",
+    "NoisyInfSplits",
+    "DiscreteNoisy13Splits",
+    "DiscreteNoisyInfSplits",
     "SPSA",
     "NoisyOnePlusOne",
     "OptimisticNoisyOnePlusOne",
@@ -319,15 +323,19 @@ def test_optimizer_families_repr() -> None:
     np.testing.assert_equal(repr(Cls()), "DifferentialEvolution()")
     np.testing.assert_equal(repr(Cls(initialization="LHS")), "DifferentialEvolution(initialization='LHS')")
     #
-    optimrs = optlib.RandomSearchMaker(sampler="cauchy")
-    np.testing.assert_equal(repr(optimrs), "RandomSearchMaker(sampler='cauchy')")
+    optim: base.ConfiguredOptimizer = optlib.RandomSearchMaker(sampler="cauchy")
+    np.testing.assert_equal(repr(optim), "RandomSearchMaker(sampler='cauchy')")
     #
-    optimso = optlib.ScipyOptimizer(method="COBYLA")
-    np.testing.assert_equal(repr(optimso), "ScipyOptimizer(method='COBYLA')")
-    assert optimso.no_parallelization
+    optim = optlib.ScipyOptimizer(method="COBYLA")
+    np.testing.assert_equal(repr(optim), "ScipyOptimizer(method='COBYLA')")
+    assert optim.no_parallelization
     #
-    optimcma = optlib.ParametrizedCMA(diagonal=True)
-    np.testing.assert_equal(repr(optimcma), "ParametrizedCMA(diagonal=True)")
+    optim = optlib.ParametrizedCMA(diagonal=True)
+    np.testing.assert_equal(repr(optim), "ParametrizedCMA(diagonal=True)")
+    #
+    optim = optlib.NoisySplit(discrete=True)
+    np.testing.assert_equal(repr(optim), "NoisySplit(discrete=True)")
+    assert optim._OptimizerClass.multivariate_optimizer is optlib.OptimisticDiscreteOnePlusOne  # type: ignore
 
 
 @pytest.mark.parametrize("name", ["PSO", "DE"])  # type: ignore
@@ -581,6 +589,14 @@ def test_shiwa_dim1() -> None:
     assert recom.value < init
 
 
+continuous_cases: tp.List[tp.Tuple[str, object, int, int, str]] = [
+    ("NGOpt", d, b, n, f"#CONTINUOUS")
+    for d in [1, 2, 10, 100, 1000]
+    for b in [2 * d, 10 * d, 100 * d]
+    for n in [1, d, 10 * d]
+]
+
+
 @pytest.mark.parametrize(  # type: ignore
     "name,param,budget,num_workers,expected",
     [
@@ -593,22 +609,39 @@ def test_shiwa_dim1() -> None:
             2,
             "DoubleFastGADiscreteOnePlusOne",
         ),
-        ("NGOpt", 1, 10, 1, "MetaModel"),
-        ("NGOpt", 1, 10, 2, "MetaModel"),
+        ("NGOpt8", ng.p.TransitionChoice(range(30), repetitions=10), 10, 2, "CMandAS2"),
+        ("NGOpt8", ng.p.TransitionChoice(range(3), repetitions=10), 10, 2, "AdaptiveDiscreteOnePlusOne"),
+        ("NGOpt13", 10, 10, 1, "Cobyla"),
+        ("NGOpt13", 10, 10, 2, "CMA"),
         (
-            "NGOpt",
+            "NGOpt13",
             ng.p.Log(lower=1, upper=1000).set_integer_casting(),
             10,
             2,
+            "HyperOpt",
+        ),
+        ("NGOpt13", ng.p.TransitionChoice(range(30), repetitions=4), 10, 2, "HyperOpt"),
+        ("NGOpt13", ng.p.TransitionChoice(range(30), repetitions=4), 10, 2, "HyperOpt"),
+        ("NGOpt13", ng.p.TransitionChoice(range(30), repetitions=10), 10, 2, "DiscreteLenglerOnePlusOne"),
+        ("NGOpt13", ng.p.TransitionChoice(range(30), repetitions=10), 10, 2, "DiscreteLenglerOnePlusOne"),
+        ("NGOpt14", 10, 601, 1, "CMA"),
+        ("NGOpt14", 10, 601, 2, "CMA"),
+        (
+            "NGOpt14",
+            ng.p.Log(lower=1, upper=1000).set_integer_casting(),
+            601,
+            2,
             "DoubleFastGADiscreteOnePlusOne",
         ),
-        ("NGOpt8", ng.p.TransitionChoice(range(30), repetitions=10), 10, 2, "CMandAS2"),
-        ("NGOpt8", ng.p.TransitionChoice(range(3), repetitions=10), 10, 2, "AdaptiveDiscreteOnePlusOne"),
-        ("NGOpt", ng.p.TransitionChoice(range(30), repetitions=10), 10, 2, "DiscreteLenglerOnePlusOne"),
-        ("NGOpt", ng.p.TransitionChoice(range(3), repetitions=10), 10, 2, "DiscreteLenglerOnePlusOne"),
+        ("NGOpt14", ng.p.TransitionChoice(range(30), repetitions=4), 601, 2, "DiscreteLenglerOnePlusOne"),
+        ("NGOpt14", ng.p.TransitionChoice(range(30), repetitions=4), 601, 2, "DiscreteLenglerOnePlusOne"),
+        ("NGOpt14", ng.p.TransitionChoice(range(30), repetitions=4), 10, 2, "MetaModel"),
+        ("NGOpt14", ng.p.TransitionChoice(range(30), repetitions=10), 10, 2, "MetaModel"),
+        ("NGOpt14", ng.p.TransitionChoice(range(30), repetitions=10), 10, 2, "MetaModel"),
         ("NGO", 1, 10, 1, "Cobyla"),
         ("NGO", 1, 10, 2, "OnePlusOne"),
-    ],  # pylint: disable=too-many-arguments
+    ]
+    + continuous_cases,  # pylint: disable=too-many-arguments
 )
 @testing.suppress_nevergrad_warnings()
 def test_ngopt_selection(
@@ -616,11 +649,22 @@ def test_ngopt_selection(
 ) -> None:
     with caplog.at_level(logging.DEBUG, logger="nevergrad.optimization.optimizerlib"):
         # pylint: disable=expression-not-assigned
-        optlib.registry[name](param, budget=budget, num_workers=num_workers).optim  # type: ignore
+        opt = optlib.registry[name](param, budget=budget, num_workers=num_workers)
+        # pylint: disable=pointless-statement
+        opt.optim  # type: ignore
         pattern = rf".*{name} selected (?P<name>\w+?) optimizer\."
         match = re.match(pattern, caplog.text.splitlines()[-1])
         assert match is not None, f"Did not detect selection in logs: {caplog.text}"
-        assert match.group("name") == expected
+        choice = match.group("name")
+        if expected != "#CONTINUOUS":
+            assert choice == expected
+        else:
+            print(f"Continuous param={param} budget={budget} workers={num_workers} --> {choice}")
+            if num_workers >= budget > 600:
+                assert choice == "MetaTuneRecentering"
+            if num_workers > 1:
+                assert choice not in ["SQP", "Cobyla"]
+        assert choice == opt._info()["sub-optim"]
 
 
 def test_bo_ordering() -> None:
@@ -652,8 +696,8 @@ def test_ngo_split_optimizer(
     budget: int,
     expected: tp.List[str],
 ) -> None:
-    param: ng.p.Parameter = ng.p.Instrumentation(
-        ng.p.Instrumentation(
+    if fake_learning:
+        param: ng.p.Parameter = ng.p.Instrumentation(
             # a log-distributed scalar between 0.001 and 1.0
             learning_rate=ng.p.Log(lower=0.001, upper=1.0),
             # an integer from 1 to 12
@@ -661,17 +705,16 @@ def test_ngo_split_optimizer(
             # either "conv" or "fc"
             architecture=ng.p.Choice(["conv", "fc"]),
         )
-        if fake_learning
-        else ng.p.Choice(["const", ng.p.Array(init=list(range(dimension)))])
-    )
+    else:
+        param = ng.p.Choice(["const", ng.p.Array(init=list(range(dimension)))])
     opt: base.OptCls = (
         xpvariants.MetaNGOpt10
         if name is None
         else (optlib.ConfSplitOptimizer(multivariate_optimizer=optlib.registry[name]))
     )
     optimizer = opt(param, budget=budget, num_workers=num_workers)
-    names = [o.optim.name if o.dimension != 1 or name is None else "monovariate" for o in optimizer.optims]  # type: ignore
-    assert names == expected
+    expected = [x if x != "monovariate" else optimizer._config.monovariate_optimizer.name for x in expected]  # type: ignore
+    assert optimizer._info()["sub-optim"] == ",".join(expected)
 
 
 @skip_win_perf  # type: ignore
@@ -771,3 +814,22 @@ def test_cma_logs(capsys: tp.Any) -> None:
     captured = capsys.readouterr()
     assert captured.out == ""
     assert captured.err == ""
+
+
+def _simple_multiobjective(x):
+    return [np.sum(x ** 2), np.sum((x - 1) ** 2)]
+
+
+def test_pymoo_pf() -> None:
+    optimizer = ng.optimizers.PymooNSGA2(parametrization=2, budget=300)
+    optimizer.parametrization.random_state.seed(12)
+    optimizer.minimize(_simple_multiobjective)
+    pf = optimizer.pareto_front()
+    fixed_points = [[0.25, 0.75], [0.75, 0.25]]
+    for fixed_point in fixed_points:
+        values = _simple_multiobjective(np.array(fixed_point))
+        # check pareto front contains a candidate dominating fixed point
+        assert any(
+            _simple_multiobjective(x.value)[0] < values[0] and _simple_multiobjective(x.value)[1] < values[1]
+            for x in pf
+        )